Delete sale offer
Delete sale offer only when owner id and sale offer are set

Envoyé depuis mon iPhone.
P.S. : Ce commit est certifié sans gluten

+++ business/services/excel.py
@@ -156,4 +156,5 @@ def clean_laboratory_sale_offers(results):
                    map(lambda r: r['result'].reference if r['result'] else None, results)
                    )
         )
-        delete_deprecated_sale_offers(succeeded_sale_offer_ref, owner_id)
+        if owner_id and succeeded_sale_offer_ref:
+            delete_deprecated_sale_offers(succeeded_sale_offer_ref, owner_id)
